RATIS-1129. Move out the RPC related APIs from DataStreamOutput.
