+++ ratis-client/src/main/java/org/apache/ratis/client/api/MessageStreamApi.java
@@ -31,11 +31,11 @@ import java.util.concurrent.CompletableFuture;
  * the leader creates a raft log entry for the request
  * and then replicates the log entry to all the followers.
  *
- * Note that this API is similar to {@link org.apache.ratis.client.RaftClient#sendAsync(Message)}
+ * Note that this API is similar to {@link AsyncApi#send(Message)}
  * except that {@link MessageStreamApi} divides a (large) message into multiple (small) sub-messages in the stream
- * but {@link org.apache.ratis.client.RaftClient#sendAsync(Message)} sends the entire message in a single RPC.
+ * but {@link AsyncApi#send(Message)} sends the entire message in a single RPC.
  * For sending large messages,
- * {@link MessageStreamApi} is more efficient than {@link org.apache.ratis.client.RaftClient#sendAsync(Message)}}.
+ * {@link MessageStreamApi} is more efficient than {@link AsyncApi#send(Message)}}.
  *
  * Note also that this API is different from {@link DataStreamApi} in the sense that
  * this API streams messages only to the leader
